Update patch set 6
Patch Set 6: Code-Review-1

(1 comment)

The task to request closing chairwiki is stalled T184961.  The last entries:

On Dec 4th 2018 RayStorm wrote:
> ...  DO NOT close the wiki. I will be getting Legal involved for their own assessment, ...
> I need to be able to access and give access, maybe even edit, which cannot happen if the wiki is closed.


We can get it closed when there is an agreement on the task. Meanwhile this change is not needed and trivial enough to restore later on.

Patch-set: 6
Reviewer: Hashar <24@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1