Update patch set 6
Patch Set 6:

> Colon removing will also be needed in commit message.

I removed it, since it's not necessary there. But be aware in the settings code,  it was addition not removal. I earlier removed it, but they said the colon is inherent part of the word, I don't know whether this will cause clash with the standard namespace partition colon.

Patch-set: 6
Reviewer: Ammarpad <6619@e9e9afe9-4712-486d-8885-f54b72dd1951>