Use excimer to set a graceful wall clock time limit in PHP 7
Throw an exception when the wall clock time limit is exceeded, instead
of using set_time_limit() which is a CPU limit.

The exception will be thrown when the VM is re-entered; this won't
interrupt a long-running C function like a MySQL query. The server-side
query killer can be used. Appropriate timeouts should be set for direct
socket operations. FPM's request_terminate_timeout can be used to kill
the worker as a last resort.

Change-Id: I8faf8f719ac8065d3968444918fb735099fcebf0