tests: Assert that no computed lists are used in wmf-config
The existing test for the naming convention was mostly a proxy
for this already, in that, it required computed lists be named
as such, and seeing "-computed" in InitialiseSettings would
probably stand out and be unexpected.

But, the problematic code isn't using it in InitialiseSettings,
the problem is loading it in CommonSettings, which isn't looked
at nearly as much. This commit adds a test explicitly for the
thing we actually care about.

Change-Id: I863ad17bb7104175310cea0e99f1656c103e28ca