Set MW_NO_SESSION for various entry points in w/
Has been enforced in core for load.php and opensearch_desc.php,
and here for static.php for several months without issue.

But, flame graphs do still show SessionManager for other entry
points that we forgot about, so let's enable it here as well.

First, use the warn stage which will allow sessions to be used
still with warnings sent to Logstash (if violated). Later,
we can set it to `1`.

Change-Id: Ia1e610a5f28585f8529cb925433abc9872e4c750