scap prep: ensure old directories exist, add logging
If we were to rebuild a deployment server from scratch, and we ran scap
prep [version] it would fail when attempting to copy old non-existant
directories. This patch checks to see if the old version directories
exist before attempting to use them.

Also, scap prep currently doesn't log very much information, so fix that
as well.

Change-Id: If7d9e312a49c2231e28e4642aa2fdda5077548b2