build: on CI only lint changed files
Running PHP CodeSniffer against all files takes a while. I wrote a
bootstrap utility for phpcs to only run it against files changed in the
proposed patchset. That is done by checking whether the environment
variable JENKINS_URL is set, and if so only pass files changed in the
patch to phpcs.

To test the behavior locally:

  JENKINS_URL=1 ./vendor/bin/phpcs -s -p .

For local environment, the default is still to process all files.  When
one of .phpcs.xml or composer.json is altered, all files are processed
as well.

Add the CI bootstrapper to phpcs.xml. CI run will be slightly faster as
a result.

Change-Id: Ic5ff9008df13cb4f3365f55d988f6dc403f99fed
Reference: T158974