Restore bureaucrat rights on hi.wiktionary to default
Bureaucrats being able to desysop at hi.wiktionary is abnormal and is contrary to the limits on configuration changes. This ability is only to be granted to very big projects where proper oversight and a poll of bureaucrats are able to mutually control each other. Currently hi.wiktionary has one inactive bureaucrat and a couple of sysops. This is also a leftover to revert from this RfC. This is also cleanup for the RFC at: https://meta.wikimedia.org/wiki/Requests_for_comment/Userrights_on_Hindi_Wikipedia

Bug: T214765
Change-Id: I411c5df56010f0386197db66a4429d2c14fc4ebc