[cirrus] decrease regex timeouts by 25%
decrease client and elastic timeouts by 25%.
With HHVM the curl request fails as a classic request timeout when the
max 60s allowed in set_time_limit are hit.
With PHP7 and ExcimerTimer we must give some time for pre&post request
processing as it may fail our request just after we receive the curl
request timeout.

50s seems reasonable in most cases. But given the nature of our API
it's possible that MW clients send 2 regexes in the same API call
(using generator=search&list=search). In which case we may hit
ExcimerTimer on the second curl request call. Ideally we'd like to
adjust the curl request timeout with the time left in the current php
request, not an absolute value.

Bug: T216860
Change-Id: If36d52646f10f9f1ba62008e2965d7cec9cb9695