WIP: mirror udp2log data into the logging pipeline
At least two problems still to be solved:

* channel routing, specifically each syslog message needs to have the channel
somewhere. Likely at the beginning of the payload so that upon receiving the
message we can detect the channel and write the rest of the payload (syslog
message) to the appropriate channel file. In other words every syslog message
needs to be formatted/prefixed with the channel name it should be routed to.

* sysloghandler level defaults to DEBUG, although we want it to use the
configured level in $opts['udp2log'] which is a string, whereas the
sysloghandler argument is an int

Bug: T126989
Change-Id: Ieeb779a0180b198dea0ccd8747f1d4409ae0154e