Deploy WikimediaEditorTasks to Beta
Configures WikimediaEditorTasks and enables it on beta wikidatawiki.

The extension is generic in principle and could be enabled on any or all
wikis, but since the feature it supports is focused on Wikidata, we only
need to enable it there at present.  Still, like ReadingLists, it will use
a shared DB/cluster in production in order to support future use cases
involving one or more other wikis.

Bug: T218137
Change-Id: Ia550485b003ddbd73b64f05b571f68f8ced0b0d8