Update patch set 1
Patch Set 1:

The choice of cluster/DB should be reviewed by DBAs but adding them on this patch without context is probably not that helpful. It would be nice to have a task with links to the DB schema, planned config and usage estimates and ask for a DBA review of that.

Patch-set: 1
Reviewer: Gergő Tisza <389@e9e9afe9-4712-486d-8885-f54b72dd1951>