Update patch set 1
Patch Set 1:

I think this is the right thing to do (maybe timeout of 2 or 3 just in case?) but I'd like to see if we can figure out what's wrong with non hasty mode with load of the type we are sending with api-request.  It should work, and will need to!  These events are safer to figure this out with than the prod eventbus ones we'll do later.

Patch-set: 1
Reviewer: Ottomata <55@e9e9afe9-4712-486d-8885-f54b72dd1951>