wikitech: Use cn:caseExactMatch: as account search filter
Since the migration of LdapAuthentication to the AuthManager framework
we have had an annoying bug in Wikitech where usernames are not being
consistently canonicalized. This can result in multiple local MediaWiki
accounts tied to the same backing LDAP entity. This is both confusing
for the end user and aggravating for debugging.

Switching the search attribute to "cn:caseExactMatch:" will tell the
LDAP directory to only consider an entry to be matching if the "cn"
attribute and the supplied wikitech username are exactly identical. This
may cause problems for a few wikitech users with lowercase "cn"
attribute values, but those can be found and corrected.

Bug: T165795
Change-Id: I80ebf0d78c65e34b295707f22ae3af4f5e34905e