speed-tests: Make static rev 872156204 more standalone
Given we already have T211487 to serve as base-line for the wiki,
the purpose of static rev 872156204 should only be to be a base-line
for the infra. That means it's important that it never changes,
not even the thumbnails (e.g. cache miss and thumbor regeneration,
WebP versions going on and off, overwrite-upload on Commons, etc.)

So I'm copying the thumbnails locally as well. This makes it a bit
less like production wiki pages, but I think that's okay, because
this isn't for absolute comparison to wiki pages. It's just as
base-line for measuring relative change on the infrastructure.

Bug: T185446
Change-Id: I1ab99b2c144fe98e52c080b4e79e8e2b982ee8da