profiler: Remove code for virtually prepended frames in excimer samples
This didn't work because the regular expression didn't match.
The first frame is "/srv/mediawiki/w/api.php", whereas the code
was expecting it to be SCRIPT_FILENAME, which for web reqs is like
"/srv/mediawiki/docroot/wikibooks.org/w/api.php".

This in turn is a symlink to "/srv/mediawiki/w/api.php", but alas,
won't match.

Rather than fixing this, remove the code.

* The frame for 'basename' (api.php) is redundant. This existed
  for two reasons:
  1. To have a generic frame at the start that doesn't vary
     by wmf branch. This is already the case thanks to the multiversion
     "/w" wrapper.
  2. To have a slash-less frame that xenon-log.py can pick up for
     grouping frames in a related file. This is not needed because
     xenon-log's regex actually does tolerate slashes and is able
     to extract the tag "api" from both "api.php" and "/w/api.php"
     without issue.

* The frame for the HTTP verb is no longer considered useful.
  And we'll drop it as part of the PHP7 migration.

Bug: T176916
Change-Id: I8c7f8c58313d227a6d9959b9f3a1c470d0656442