db-eqiad.php: Depool all slaves in x1
I would like to depool all slaves in x1 and leave db1069 serving reads
to test an schema change done with replication directly on the master.
The schema change is on the echo_notification table, which is small in
most of the wikis, the biggest one is enwiki (18GB).
The schema changes are mostly index related and just two column drops.
My idea is to test a few of the first wikis (really small ones) and see
how it goes.
Also check what is the impact on leaving the master serve reads.

Bug: T143763
Change-Id: I54888a001b11fff63b6ab9b69c054b6fde2c0ab4