Train: scap clean, feature flag prune branches
Gerrit's use of HTTP auth tokens has been temporarily removed. Until it
is re-enabled we rely on Gerrit's SSH auth. In the case of scap clean,
since scap overrides the SSH_AUTH_SOCK env var with a pointer to
keyholder, we are not authorized to prune any branches on Gerrit since
there are no appropriate keys in keyholder.

Until either Gerrit's HTTP auth tokens are re-enabled, or we have a
shared key to prune branches in keyholder, we need to feature flag
branch prune so that we ensure that we're doing the rest of scap clean
(removing old branches on the application servers themselves).

Bug: T218783
Change-Id: Ie967ddf3ff2772f840bc8e2bf20ae99e29aeae3f