Fix wgImportSources setting for wikidata dblist
The old code set the setting for the wikidata dblist (containing
wikidatawiki and testwikidatawiki) and then again for testwikidatawiki
directly (to the same value). It seems clear to me that the first entry
was supposed to be wikidatawiki, not wikidata, but for this setting I
think it makes more sense to actually use the wikidata dblist and group
it with the other generic project entries, wikisource and wikiversity.

Change-Id: If62b77e3e100113cca3923cb4fbf01e9181d4451