profiler: Increase max stack depth for sampling profiler to 250
For HHVM/Xenon, a 24h window of samples (2019-03-26) contained
only stacks with a depth upto 229.

Assuming the cost of serialising this depth is acceptable,
250 would be our upper bound after which we can afford to
discard stacks, given they're not known to happen in prod under
normal circumstances.

This commit does not do the discarding part, and does not resolve
the issue of it being possible for ArcLamp to receive incomplete
stacks. See task for more details on that.

Bug: T176916
Change-Id: I4e12ea320e68920af3c1b3b36dbd66b73ec6399b