Add a new Apple Touch icon to configuration
This patch adds the new Apple Touch icon for the German Wiktionary
(uploaded as I027d345) to InitialiseSettings.php so that it can be linked
from the <head /> tag instead of the default logo used by other Wiktionary
projects.

See task description for a link showing community consensus for the change.

Bug: T202902
Change-Id: Ia519907f469f5a18b7cc5f097dff48daef90feaf