Update patch set 1
Patch Set 1:

Hm, even if that wasn’t the problem on idwiki, this still looks like a good idea to me? Or is Wikibase already supposed to check if mapframes/maplinks are enabled before emitting them, and doing it in the config was redundant?

Patch-set: 1
Reviewer: Lucas Werkmeister (WMDE) <4699@e9e9afe9-4712-486d-8885-f54b72dd1951>