Update patch set 5
Patch Set 5:

Tried it out on mwdebug1001 and tested it once by adding a throw statement, and once by replacing the substr_count check with true.

Then made a bunch of requests to that that the half a dozen sub properties in Graphite get lazy-created, which is now done.

https://graphite.wikimedia.org/render/?width=800&height=400&target=MediaWiki.arclamp_client_error.*.rate&target=MediaWiki.arclamp_client_discarded.*.rate

https://grafana.wikimedia.org/d/yVf-D1RWk/arc-lamp?orgId=1

Patch-set: 5
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>