Update patch set 1
Patch Set 1:

I don’t see this logo used on wikimaniawiki page content yet (e. g. About [1] still links to the old logo). Is there a source for this being the new logo?

Also, on Commons this file [1] seems to be licensed under CC BY-SA 4.0. I don’t know the license rules for site logos – does attribution need to be given somewhere? Or perhaps the image creator (in this case Blanca Flores) gives their permission for the image to used like this?

[1]: https://wikimania.wikimedia.org/wiki/About
[2]: https://commons.wikimedia.org/wiki/File:Wikimania2019_logo.svg

Patch-set: 1
CC: Lucas Werkmeister (WMDE) <4699@e9e9afe9-4712-486d-8885-f54b72dd1951>