Rename UploadWizard depicts/statements toggle
'depicts' was renamed to the more generic 'statements', because
other non-commons wikis (or campaigns) may not care about depicts
and configure things with other/additional default statements.

'depicts' still works, but it'd be nice to be able to remove the
backward compatibilith some day.

Also no longer only setting the value when true, but always
assigning it - this way, it doesn't matter what the default is
in UW.

Change-Id: I9b164eebfdaa6a5508ad7731974ae2fcea1435a1