Revert "Disable MessageBlobStore::clear() via hook"
This reverts commit 7f1a3bc742edf05f57cb02c295dab276a5d00a26.

Was originally disabled for performance reasons, but those reasons no
longer apply since MessageBlobStore no longer uses the database for
storage. Nowadays, MessageBlobStore::clear() is just a cheap
touchCheckKey() call.

This is not enough to fix T217719, because this runs at the wrong time
in production (see task for details), but we should at least clean this up.

Bug: T217719
Change-Id: I4a615f1e13fc7c2e76fe142411b3b0cb002e8f22