Set wgLocalisationCacheConf['storeClass'] explicitly
Instead of relying on the default 'store'=>'detect' and it picking
the CDB store, configure this explicitly.

Reason being that we use manualRecache, which means dynamic
building of l10n cache is disabled in production. As such, MW
should not *dynamically* determine which format to use because
it couldn't build it even if it was different or missing for
any reason.

Also remove the comment from 2011 because 1) it is no longer true
because MediaWiki has been using $wgCacheDirectory as default for
this option for several years, and 2) using old-style l10n-caching
would likely be a very bad idea in 2019.

Bug: T99740
Change-Id: I576f45f5de79a93dafb3e2b111714b7a5845ce93