Update patch set 1
Patch Set 1:

Merged now it will be a no-op since the code changes for this are not ready/deployed. I'm creating this to discuss the approach.

The idea of this is to be able to configure event destination per-wiki per-stream. The change in InitializeSettings allows us to create a mapping between the topic/stream name to the destination name. The change in CommonSettings allows us to configure a custom EventFactory per event service name. If the EventFactory is not intended to be used (eventgate-analytics), the factory can be not configured. Different factories will share 99% of the code, we will just override 2 methods: createEvent (to set different meta, stream, $schema etc.) and a new public method 'getStream' (name to be discussed) so that we can fetch the '.meta.stream' from eventgate events and '.meta.topic' from eventbus event.

This is all clearly overcomlicated and will not really be needed after transition, so I guess we will remove it after. Thought?

Patch-set: 1
Reviewer: Ppchelko <2681@e9e9afe9-4712-486d-8885-f54b72dd1951>