Update patch set 1
Patch Set 1:

So the idea is to first build the event with the stream name, and then have getStream pull that out of the event, and use the returned value to route to the configured EventServiceName?

I like.  

As for the configurable EventFactory, dunno.  I think we should keep the routing code after we finish this migration, but for now just put temporary conditionals in the code for the varying ways to build meta.stream vs. meta.topic, and $schema vs meta.schema_uri.  Once all events are using the new format, we just remove the code for the old format.

Patch-set: 1
Reviewer: Ottomata <55@e9e9afe9-4712-486d-8885-f54b72dd1951>