Update patch set 1
Patch Set 1:

Hm... I think I want to abandon this change. This idea basically makes the process of what to construct and where to send it cyclic - we need a topic name to get an appropriate instance of EventBus which will give us the event we will take the topic name out of. This is solvable either by duplicating the topic names or by a lot of moving the code around inside the Eventbus extension, but since we will not need this configuration after the transition I believe we should do it. Let's just go with code deploys?

Patch-set: 1
Reviewer: Ppchelko <2681@e9e9afe9-4712-486d-8885-f54b72dd1951>