Allow MW to honour the X-Request-Id header if set
As of I605471fb8b5bbc290baeecc7d80d9d715cb240c9, MediaWiki can honour
the externally-set `X-Request-Id` header and use is the request id for
its own operations if `$wgAllowExternalReqID` is set to `true`. This is
safe to activate in production because Varnish unsets the header for all
incoming external requests.

Bug: T201409
Change-Id: Id3bdec6e8c9158a34881cc5c3d67a75cf09e3fce