Simplify VisualEditor config variables (1/2)
The config variables $wmgVisualEditorSingleEditTabSecondaryEditor and
$wmgVisualEditorSecondaryTabs had a very similar purpose of describing
whether the visual or the wikitext editors was the "primary" option.
One only applied in SET mode and the other only in non-SET mode. Now
there is just one variable $wmgVisualEditorIsSecondaryEditor.

The new variable will also control which editor opens upon clicking a
red link on non-SET wikis (T223793). But right now this is a no-op.

Bug: T223793
Change-Id: I45235af6a9eb61413d34595087722eeff7c4304c