Remove bureaucrat protection level for all Serbian projects
To protect or unprotect a page, only sysop permissions are required.
Because of that, protection levels requiring higher level of permissions
than something a sysop normally has doesn't make sense.

In T215653, a patch that allows sysops to protect a page to be edited
only by bureaucrats was uploaded. Before this level of protection can
be implemented in a working way, it should be discussed if something like
this should be allowed to work, and if it doesn't have some negative effects.

Bug: T217005
Change-Id: I66e2140d24ecf459837673d9e4ce07bc897118ea