Update patch set 2
Patch Set 2:

I've moved the hhvm equivalent to puppet recently, in prep for making it share the error-page.erb template. Might make sense to do this for the php7 one right away? I assume it will be referenced from an INI setting declared by puppet, which could then also control where the file is provisioned. For HHVM this previously relied on the wmf-config repo being cloned in a particular location and having the file which seemed suboptimal.

Patch-set: 2
CC: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>