Update patch set 3
Patch Set 3: Code-Review+1

This is ok to me as is, just some wondering- should we reduce db1081 load in advance? I think the last connection issue we had is to pool immediately the old master, not to depool a lot of weight quickly, but I wanted to mention it.  I don't think it should be a huge issue anyway.

Patch-set: 3
Reviewer: Jcrespo <2556@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1