Configuration migration for Translate
This is part of a soft migration for extension registration in Translate.
For sites where translation memory is disabled and the list of translation
memory services is empty (only otrs_wikiwiki), the default service setting
is ignored.  We need this new setting in order to deploy Ica4beddcaa95946,
to distinguish the disabled translation memory from the fallback case
which will enable built-in TTMServer configuration.

Bug: T87985
Change-Id: I590c7841be7044b2e57d1cdda5c6065c8236fad1