Update patch set 2
Patch Set 2:

This won't work, because the StructuredDiscussions code has been removed from enwiki (even though it's meant to be live on all prod wikis).

Instead of working, it errors with "Feedback report failed because MessagePoster could not be fetched". See the config for dewikivoyage etc.

Obviously we can re-enable the extension on enwiki, but I imagine that might confuse some people.

Patch-set: 2
CC: Jforrester <366@e9e9afe9-4712-486d-8885-f54b72dd1951>