Specify CentralAuth and OAuth session storage separately from per-wiki session storage.
As part of the Kask rollout, it is necessary to separate the
session storage used by CentralAuth and OAuth from per-wiki
session storage.
This commit specifies CentralAuth and OAuth session storage
separately, but makes no change in the actual storage location. Per-wiki session storage will be repointed to Kask in separate commits.

Bug: T227097
Bug: T227696
Change-Id: Ibf1942a9ebf7268abc3462565f0c6f00e2e2124a