Remove /w/skin-1.5 symlink
The time has come for this symlink to die. Any file that existed
under this path when it wasn't deprecated has been removed or
renamed at least once, thus all those referencces are bound to
be 404 Not Found already (or even not resolve entirely, as a good
number of them are actually through bits.wikimedia.org which no
longer exists).

In MediaWiki Codesearch (Everywhere), only 1 result, from the
DumpHTML extension. The url used there is already broken,
and also that extension is no longer deployed at WMF.

In mwgrep (on-wiki site scripts and gadgets), only 2 results
(foundationwiki and wikimania2007wiki), both in CSS comments,
not actual code.

Bug: T156319
Change-Id: I42dd7e1839745ed7b300cd1107f934a0644f476e