Specify $wmgWBRepoConceptBaseUri again
This is preparation for specifying $wgWBRepoSettings['conceptBaseUri'] –
the default value in Wikibase is correct, but computed using a callback,
which produces a debug log warning on each request. Let’s avoid this and
explicitly configure the base URI for Wikidata and Test Wikidata.

$wgWBRepoSettings['conceptBaseUri'] will actually be set in the next
commit, to ensure that this part of the change is merged and synced
first – if both parts were one commit, and Wikibase.php was synced
first, clients would see errors due to $wmgWBRepoConceptBaseUri not
being set.

This also fixes the issue that caused the revert of commit 49cc5176e1
(change I8f6a6d67d7) – commonswiki and testcommonswiki also need a
$wmgWBRepoConceptBaseUri.

Bug: T225212
Change-Id: Ib3584f456e2a4810c3e6951453e8b382bf7c9557