Define settings for Citoid+Wikibase integration
We first define the settings and then later use them in Wikibase.php so
that the files are guaranteed to be synced in the right order (syncing
Wikibase.php first would result in errors).

For now, the settings correspond to their default values, so there
should be no change even once they are used in Wikibase.php.

Bug: T228414
Change-Id: Ia7e3191299d165217b035cc3d0230e10974d46b0