Switch "wait for replica" method to use GTIDs for external DB clusters
For the main wiki DBs, MW was already configured to use GTIDs as the
method behind "wait for replica" and pt-heartbeat as the method behind
"get lag times".

But, the MW core default of binlogs for "wait for replica" and
Seconds_Behind_Master for "get lag times" was still implicitly used for
external DB clusters; these are connections to extension1 (e.g. Flow),
and ExternalStore (e.g. wikitext blobs). [1]

* Enable GTIDs everywhere, so that external DBs now use those instead
  of binlogs when doing "wait for replica", matching what we did for
  the main wiki DBs already.

* Keep avoiding use of pt-heartbeat on external DBs, but
  make that more explicit (avoid code duplication).

[1] https://wikitech.wikimedia.org/wiki/MariaDB#Extension_storage

Change-Id: I46d2b811f62876897aeb7e6423f9cb8d6b27fac4