scap: prep and clean git ops for /srv/patches
Part of "scap prep" is to copy patches from one version to the next.
Part of "scap clean" is to remove patches of old versions. Since the
patch directory is a git repo these actions leave a mess behind.

This change runs (effectively):

    git -C /srv/patches add --all
    git -C /srv/patches commit --quiet -m 'Scap [action] for "[version]"'

For each action scap takes with that repo.

Bug: T222240
Change-Id: I8770d794636f1c5ca7b51907475bbfb3f0202fd8