Switch testwiki to read sessions from kask, with fallback to redis
Stage one of using testwiki as a testing server for kask session
storage was to write to both redis and kask, with reads from
redis (primary) and kask (fallback). That worked well, so this
second stage is to still write to both, but read from kask
(primary) with reads from redis (fallback).

Bug: T222099
Change-Id: Ia8450f5d645af797f37814654680cd039914e70e