Remove related-articles-footer-blacklisted-skins.dblist
Wiki list for which we disable the related articles footer on all
skins, regardless of other configuration. Wasn't actually doing
anything to begin with, and it looks like we don't have any particular
reason to disable the extension on Minerva, the default; if people
weren't complaining before, no need to fix this for that.

(Now it's just causing RA to appear on all skins, though, which they
are complaining about, and we probably should fix.)

Bug: T229644
Change-Id: Ia167c7ff028856ad70b52cf679f5249599e5a9ea