Use AbuseFilterParser on beta enwiki
The beta cluster uses AbuseFilterCachingParser because it is buggy, and this way we can find new errors. However, this makes it impossible to test parser-related changes on the beta cluster, at least for the non-caching parser. This patch enables the classic parser on deploymentwiki only, as an arbitrary choice to have a place on beta cluster with the old parser.

Change-Id: I1748bb2bef71e34ad3e0710e6351771c2cc2d6e5