General cleanup of `groupOverrides`.
Entries that just reaffirm the `groupOverrides2` defaults, or, if those
don't specify anything regarding a specific assignment, entries that
reaffirm mediawiki's default settings, are extraneous. Additionally,
user groups that recieve three or fewer rights can generally be written
in one line, reducing the size of the file and improving the readability
of the code.

Bug: T231041
Change-Id: Ibc159f5f5e6698cabc4454e3d76299ec6c2901a1