Drop MobileWebUIActionsTracking sampling rate to 0.01%
We're going to track lots of new actions with the new
MobileWebUIActionsTracking schema and we expect a major
bump in the amount of events per second.

To play safe we're starting from very safe 0.01% sampling rate
and once code gets pushed to all wikipedias we will slowly bump
the sampling rate.

Bug: T220016
Change-Id: Iacfc27fd84bdcf6af42497260c5480b79eb1e121