beta: Remove unused wgConf feature of '+beta' tag
wgConf has a notion of tags to apply a setting to multiple wikis,
e.g. 'wikipedia', 'special', and 'fishbowl'.

In InitialiseSettings-labs.php the code registered an extra beta-only
tag called 'beta' that is set for all wikis, which explains why
it wasn't used because this file is only for settings that apply
to beta. Setting 'wgSitename' => [ 'beta' => 'Beta' ] would be no
different from '-wgSitename' => [ 'default' => Beta' ], except
that it would allow wiki-specific overrides to still exist.

Nonetheless, it was only used in one place in InitialiseSettings-labs.php.
Not anywhere else or in InitialiseSettings.php. I've removed the
feature and replaced the one use with a direct assignment.

Ref T223602 as this reduces reliance on wgConf and brings us one step
closer to using a static array.

Bug: T223602
Change-Id: I9a752eb9f462d2e655d1731310ad2884c890a24f