beta: Remove redundant 'flow_only_labs.dblist'
Use flow-labs.dblist and break inheritance with prod IS.php
for this array key. The db lists are maintained separately
for better or worse. The assumption that Beta would set
all prod-equiv wikis to 'true' and then set 0 or more extra
onees to 'true' as well seemed rather fragile. I'd expect that
if a wiki is in 'flow.dblist' but not 'flow-labs.dblist' that that
means it is only enabled in prod. This change makes that true.

Change-Id: I5d75d292c90ee619d38cd3da63f9381455a9f05d