beta: Remove use of SiteConfiguration->siteParamsCallback feature
This was only used by Beta and made it quite confusing to figure
out how these tags get computed. Basically, when MWConfigCacheGenerator
calls wgConfig->getAll(), the SiteConfiguration class looks if its
public 'siteParamsCallback' was overridden, and if so, call that
to get additional params/tags - merging them with the ones passed
to getAll() directly.

There's two ways to avoid this. One, is by forking getMWConfigForCacheing
into a prod and beta version and adding a conditional in CommonSettings.php
where we call it. Or to pass it in and handle explicitly here from
a single, testable, method.

Change-Id: I36a8cdac382a1fc02fe8a88ef7d0995b6f23c339