beta: Unbreak config vars that used '$variant' placeholder
This isn't a real PHP variable, it was substituted by wgConf
until I broke it in d495d5e249499.

Fix by using a literal instead. We can add this back in in the
future, if needed, to vary all of '.beta.wmflabs.org' like we
do for some service and puppet configs already.

Change-Id: I5e0e2a509ff31835da87949f25643affd089e3a3