tests: Remove dbconfigTest.php
This had tests for wgLBFactoryConf['sectionLoads'] which no longer
exists in db-*.php files.

This issue was obscured by the fact that the file also referenced
$wgSecretKey which wasn't mocked. Whilst PHPUnit was trying to report
this error, a bug in PHPUnit 6 causes it to *first* check for
any diry state in GLOBALS.

Where it will find "problems" with wmfRealm and wmfDatacenter
being mocked but not yet restored, which is obvious because we
haven't reached restoreGlobals() yet, which was the next line
in DbconfigTest::loadDbFile() after inclusion of db-eqiad.php.

The only reason it is checking for globals in the middle of the
test is because of the $wgSecretKey undefined variable, but
once it has found the dirty globals, PHPUnit 6 aborts the test
and complains only about that.

* Remove all this, as it is no longer needed.
* Simplify WgConfTestCase while at it, and en-anble
  the Exception from teardown.

Bug: T232691
Change-Id: If5f3ccd86d203ebde4c3234238a43585ea07b7bd