Uninstall dev deps from production vendor/
Effectively undoes b88a57c4ce9c62, and re-applies 17a7f7fc30abf.

Change b88a57c4ce9c62 was merged on 12 Sept. Based on SAL, it was
deployed on 17 Sept by an unrelated scap for 1.34.0-wmf.23.

These are not needed in production MediaWiki, and have not yet been audited
for security, and in general impact on it here (wrt repo size).

There is also the potential for version clashes with MW deps.
This is important because we currently run zero tests in CI on
MediaWiki with wmf-config installed.

At least one library (composer/semver) appears to clash with
MediaWiki currently, which means its version was now determined by
wmf-config and the MediaWiki one shadowed/ignored. We used to have
a similar issue with wikimedia/cdb in the past.

Including dev deps in prod can certainly be considered, but
should imho follow a wider discussion about how, why and possibly
other solutions to the problem(s) at hand.

Change-Id: I9e47d30b528f4233c5aaa58a747b7db5503416ce