Variant configuration: Allow for YAML-based inheritance of configuration
This will allow for us to move the contents of InitialiseSettings into the YAML
files in a manner that provides proof that the output is identical.

The initial use is to replicate the DB lists in the YAML files, so that they
can be calculated automatically at repo build time instead of via a Makefile.

Further uses are in a series of follow-ups for clarity.

Bug: T223602
Change-Id: I4ce08a95f557910253087974e992ab32555b53aa