Enforce PHPUnit global leak strict and fix StaticSettingsTest leak
Turns out, upstream doesn't want the "beStrictAboutChangesToGlobalState"
option to ... be strict about changes to global state.

Rather, it just detects it and prints a forgettable warning.
To actually be strict you need failOnRisky="true", which gets
us there.

Without the fix in StaticSettingsTest, the tests fail with
this commit applied.

I also found the sourc of the leak, it's the global assignment
from MWRealm.php.

Bug: T233771
Change-Id: I0e1fdf8979a44672e6f3420cae59f85d3e750360