[cirrus] Disable instant indexing on wikidata
... and reset to the default 30s refresh interval.
The reason is that wikidata completion API is now using the DB in
conjunction with elasticsearch.  This was done in
Icfa9e1d88dc7185bec7c4b5346e50b87a62692f8 to address T206256.  We
should no longer need to enable this feature which causes the entities
and properties to be sent twice to the elasticsearch backend.

Change-Id: Ifb5ca8ca7c789fee0db4772c6698f9b2a9c43dd8