static.php: Explicitly disallow dotfiles
There aren't any here that we're worried about. Anything stored
on disk under /srv/mediawiki/php-*/ should be considered public
for all intents and purposes (we store config outside of that for
this reason). But extra hardening can't hurt :)

Also generalise more of the 4xx errors toward 404, and document
this decision.

Bug: T204186
Change-Id: Icec2e956871c49d0262d63759a94720f841766e3