Migrate to Kask for Echo seen-time storage
A typical phased rollout will not work for wikis that use global
notifications.  Configuring such a wiki to use echostore while the
default remains Redis, will create a split-brain scenario with changes
to the seen time of one not reflected on others.  This changeset
updates the default to use Kask/echostore, migrating all at once.

Bug: T222851
Change-Id: I6113d3be12266f3c8d117ca1a737557da9c6cd42