Wikibase (beta-only): Update wmgWikibaseClientDataBridgeHrefRegExp
With Wikibase change I6156cc1645, the RegExp should have three capturing
groups rather than two, to capture the entity title. For (Beta)
Wikidata, that’s the same as the entity ID (we only match items, which
are in the main namespace), so it’s a copy of the same group.

Bug: T238918
Change-Id: I3ffaca712d32775000af0068fecfa85e89d568b5