Make special *.wikipedia.org sites no longer "special"
The arcbcom, and test wikis were both in the
"wikipedia" family and the "special" family.

This doesn't particularly make sense as "special" is mainly
the catch-all group for wikis that are in none of the "real"
families. That is, we place a wiki in "special" if it should *not*
inherit settings like `'server = $lang.wikipedia.org`, or if it
should *not* be included in Puppet-ised maintenance jobs that
run over each wiki in wikipedia.dblist.

Given these wikis were in both, they did inherit those settings,
and did get included in those dblist-based jobs.

There are two other things to consider, though:

1. Configuration settings that apply only to 'special', but
   not 'wikipedia'.

   There is one (1) configuration setting that does it,
   'wgLanguageCode' in InitialiseSettings.php. There the default
   is the subdomain-derived '$lang' value, but 'special' wikis
   are set to 'en'. However all the wikis this commit affects
   already set their own language code directly, and thus did
   not rely on inheriting this value.

2. Configuration settings that apply to both 'special' and
   one of 'wikipedia'.

   There are no such settings, fortunately.

Bug: T239301
Change-Id: I34b3466f605489da5b6b9006149deee4eaf90ce8