GrowthExperiments: Begin "initiation test" for suggested edits
This changes the homepage from a 50/50 on/off experiment to a
40/40/20 A/B/C experiment, where A is the standard homepage,
B is the homepage with suggested edits pre-initiated, and
C is no homepage at all.

This is implemented by setting the percentage for which the homepage is
enabled to 80%, and setting the percentage of those accounts for which
suggested edits is pre-initiated to 50% (50% of 80% = 40%).

Bug: T238888
Change-Id: I6030c6e744ed58fcb9ba5fd92982c1e7485f52dc