Make ge.wikimedia.org site no longer "special"
This was both in the wikimedia "chapter" family and the "special" family.

This doesn't particularly make sense as "special" is mainly
the catch-all group for wikis that are in none of the "real"
families. That is, we place a wiki in "special" if it should *not*
inherit settings from one of the families, or if it should *not*
get included in a cronjob that loops over all wikis in wikimania.dblist.

Given these wikis were in both, they did inherit those settings,
and did get included in those dblist-based jobs.

There are two other things to consider, though:

1. Configuration settings that apply only to 'special', but
   not 'wikimedia'.

   There are no such settings, fortunately.

2. Configuration settings that apply to both 'special' and
   one of 'wikimedia'.

   There is one (1) configuration settting that does this,
   'wmgULSPosition'. It is set for both 'special' and 'wikimedia',
   which could affect ge.wikimedia.org depending on which setting
   it currently gets randomly assinged given wgConf doesn't
   expect multiple matches and only one can win.
   Fortunately, while wmgULSPosition is set twice, they both
   set it to the same value, so this is also unaffected.

Bug: T239301
Change-Id: I278caa0a84d10ef41d57860d4d873011bb7e532f