Disable sanity check cirrus jobs for Wikidata
Currently, it puts really high pressure on the database and nodes
by trying to render all items.

Let's disable this until either we fix this job to not render items
or the migration for the new term store is done so the database doesn't
get overloaded by the volume of the queries.

Bug: T239931
Change-Id: I866f4102129e9e2574f3bc4a292546d7df583869