CommonSettings.php: Remove CLI 'display_errors=stderr' setting
This was only set in CLI mode, but no longer works there.

  $ mwscript eval.php --wiki testwiki
  > $testing = [];
  > echo $testing['not-set'];

This does not print "Notice: Undefined index: not-set" as one
might expect. The error is internally handled by MediaWiki
and sent to Logstash. It does not end up in stdout or stderr.

This followsup cd0c57975f and T138291 here it was added to handle
this example:

  $ echo 'error_reporting(-1); echo $testing;' | mwscript maintenance/eval.php aawiki 2>/dev/null

Where ` 2>/dev/null` is meant to hide the error. But, even without
` 2>/dev/null` there is no error, because it doesn't get printed.
Presumably this changed in HHVM->PHP7.

Change-Id: I30a00c5071076659db27acaf5875f3ccdabb5420