CommonSettings.php: Remove very old 'error_append_string' INI override
This traces back to the 2012 "initial commit" for the mediawiki-config
Git repository (where we imported the final state of the private SVN
repo local the deployment host).

I've tested this patch on mwdebug1001 and triggered a PHP error before
and after, confirmed the value was indeed set before and no longer set
after, and that the Logstash/syslog utilities work either way. It
does not appear to be used for anything currently.

Change-Id: I0c502976cc0ef677c77d0123c077fc88b70f2d96