scap: Clean up unused build configuration
There are no tests, so stop running nosetests (which is deprecated
anyways). sphinx is commented out, and there's no documentation source,
so remove that as well.

Since we just need to run flake8 now, use tox-wikimedia to simplify it
all. We still need usedevelop=true, because the setup.py needs to read
from requirements.txt.

Change-Id: I6b82df8aaedc529cacbe6678d979874d3fe3d5b7