Fix typo in the config name
Merging this chain will result in wikidatawiki and testwikidata wiki
reading from the new terms store for some items. This means we would
expect to see an increase in new term store access and a decrease in
old term store access. This difference may be small due to the very
limited number of items accessed by wikidatawiki and testwikidatawiki
as clients.

Bug: T244697
Change-Id: Iad05762183f9639a54b30bf6f49856b714766809