Allow non-autoconfirmed users to propose OAuth apps
New developers (and sometimes experienced developers who had little
cross-wiki activity before) are regularly bitten by the fact that
you need to have an autoconfirmed account on Meta to be able to
propose a consumer. There is little danger in removing this
limitation: owner-only consumers are basically just an alternative
login method, and non-owner-only consumers need admin review anyway.
The worst that can happen is spam, and the consumer registration
form seems like an unlikely spam target (and if someone does target
it, the autoconfirmed requirement won't help too much).

Bug: T213760
Change-Id: Ifac747ead38cd77a112ed103017f140247b13fd5